requirements.txt
requests
pandas
openpyxl
sqlalchemy
python-dotenv
